Add declinaison in headings (optional) in real boat mode regroup all A360(), diffAngle(), etc in Util A small change in interpolation ;)
git-svn-id: svn://svn.code.sf.net/p/qtvlm/code/trunk@1161 2387b160-4621-4e8a-be36-40ea10d72da2

+++ src/interpolation.cpp
@@ -604,8 +604,8 @@ void interpolation::get_wind_info_latlong_hybrid_compute(double longitude,  doub
     ratioLatDebug=d_lat - floor(d_lat);
     ratioLonDebug=d_long - floor(d_long);
 #endif
-    int i0 = floor ((longitude-gridOriginLon)/lon_step);
-    int j0 = floor ((latitude-gridOriginLat)/lat_step);
+    int i0 = (int) floor ((longitude-gridOriginLon)/lon_step);
+    int j0 = (int) floor ((latitude-gridOriginLat)/lat_step);
     if(((latitude-gridOriginLat)/lat_step)-j0!=0.0)
     {
         if(lat_step<0)
